requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-js-asset==1.2.2
django-mptt==0.9.1
django-querycount==0.7.0
django-redis==4.10.0
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
idna==2.8
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
